- Refactor Account model to include a balance field and update AccountDto, AccountControllerTest, and AccountService accordingly. - Refactor AccountService to use Lombok and improve code readability. - Refactor AccountServiceTest class and import required classes in DatabaseLoader.java. - Refactor DatabaseLoader class and add new functionality for generating TicketTypes in a concise format. - Refactor PassengerController to include additional services and endpoints for ticket and transaction manipulation and retrieval. - Add endpoints for buying tickets and adding credit to a passenger's account, along with corresponding tests. - Activate expired tickets and notify users in a scheduled task. - Add ticket type enum and related fields to TicketType entity, and implement getAllTicketTypes method in TicketTypeService to return all ticket types with their details.

+++ src/test/java/com/electronicTicket/models/ModelsTest.java
@@ -1,7 +1,9 @@
 package com.electronicTicket.models;
 
+import com.electronicTicket.models.enums.Role;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -42,11 +44,11 @@ class ModelsTest {
         TicketType ticketType = new TicketType();
         ticketType.setTicketTypeId(1L);
         ticketType.setName("Adult");
-        ticketType.setPrice(2.5);
+        ticketType.setPrice(BigDecimal.valueOf(2.5));
 
         assertEquals(1L, ticketType.getTicketTypeId());
         assertEquals("Adult", ticketType.getName());
-        assertEquals(2.5, ticketType.getPrice());
+        assertEquals(BigDecimal.valueOf(2.5), ticketType.getPrice());
     }
 
     @Test
